Quip_Auto/quip2deck/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "quip2deck"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
  "fastapi==0.114.2",
  "uvicorn==0.30.6",
  "beautifulsoup4==4.12.3",
  "lxml==5.3.0",
  "python-pptx==0.6.23",
  "pydantic==2.9.2",
  "typer==0.12.5",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]